requirements.txt
wikipedia
discord.py